+++ ta4j-core/src/test/java/org/ta4j/core/rules/IsRisingRuleTest.java
@@ -30,7 +30,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.BaseBarSeriesBuilder;
-import org.ta4j.core.indicators.helpers.FixedDecimalIndicator;
+import org.ta4j.core.indicators.helpers.FixedNumIndicator;
 
 public class IsRisingRuleTest {
 
@@ -39,7 +39,7 @@ public class IsRisingRuleTest {
     @Before
     public void setUp() {
         BarSeries series = new BaseBarSeriesBuilder().build();
-        var indicator = new FixedDecimalIndicator(series, 1, 2, 3, 4, 5, 6, 0, 1, 2, 3);
+        var indicator = new FixedNumIndicator(series, 1, 2, 3, 4, 5, 6, 0, 1, 2, 3);
         rule = new IsRisingRule(indicator, 3);
     }